% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.haps}
\alias{dolph.haps}
\title{Dolphin mtDNA Haplotype Sequences}
\format{A list of 33 aligned sequences}
\usage{
data(dolph.haps)
}
\description{
Dolphin mtDNA Haplotype Sequences
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean.
  Marine Mammal Science 31:1-20
}
\keyword{datasets}

