% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{data}
\name{dolph.msats}
\alias{dolph.msats}
\title{Dolphin Microsatellite Genotypes}
\format{A data.frame of 126 samples and 31 variables}
\usage{
data(dolph.msats)
}
\description{
Genotypes for 15 microsatellite loci
}
\references{
Lowther-Thieleking J.L., F.I. Archer, A.R. Lang, and
  D.W. Weller. 2015. Genetic variation of coastal and offshore bottlenose
  dolphins, Tursiops truncatus, in the eastern North Pacific Ocean.
  Marine Mammal Science 31:1-20
}
\keyword{datasets}

