% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Hstats.R, R/popStructStat.R, R/statChi2.R, R/statFst.R, R/statGst.R, R/statJostD.R, R/statPhist.R
\name{Hstats}
\alias{Hstats}
\alias{popStructStat}
\alias{statChi2}
\alias{statFst}
\alias{statFstPrime}
\alias{statGst}
\alias{statGstDblPrime}
\alias{statGstPrime}
\alias{statJostD}
\alias{statPhist}
\title{Population Structure statistics}
\usage{
Hstats(g, strata = NULL)

statChi2(g, strata = NULL, ...)

statFst(g, strata = NULL, ...)

statFstPrime(g, strata = NULL, ...)

statGst(g, strata = NULL, ...)

statGstPrime(g, strata = NULL, prime.type = c("nei", "hedrick"), ...)

statGstDblPrime(g, strata = NULL, ...)

statJostD(g, strata = NULL, ...)

statPhist(g, strata = NULL, hap.dist = NULL, model = "K80",
  gamma = FALSE, pairwise.deletion = TRUE, ...)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{strata}{an optional alternative stratification vector to use in place
of the stratification in \code{g}. If \code{NULL} stratification is taken
from \code{g}.}

\item{...}{optional arguments passed to or from other functions.}

\item{prime.type}{type of G'st to calculate. Can be "nei" or "hedrick".}

\item{hap.dist}{matrix of pairwise distances between haplotypes.}

\item{model,gamma,pairwise.deletion}{parameters passed to
\code{\link[ape]{dist.dna}}.}
}
\value{
the named statistic estimate.
}
\description{
Population Structure statistics
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}

