% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strataG-package.R
\docType{package}
\name{strataG-package}
\alias{strataG-package}
\alias{strataGdevel}
\title{Summaries and population structure analyses of DNA sequence genotypic
  data}
\description{
strataGdevel
}
\keyword{package}

